W8_A2/requirements.txt
google-generativeai
pypdf
python-dotenv